Add keep_artifact option to be able to remove the source AMI after copy
This is a feature that give the choice whether to keep the source AMI
or not after copy.

It solves the use case where we have to copy an AMI that is using
encrypted snapshots across accounts.

We can only share snapshots encrypted with Customer Managed Keys, but
the resulting copied AMI is using the default `aws/ebs` KMS key.

For consistency across all accounts we can then use the post processor to copy the AMI on the source account itself then drop the intermediate AMI that used the CMK.