Implement a 'tags only' mode
I'm quite comfortable with EC2's "shared AMI" model, where you only pay for
the storage in one place, but I do like to have my images well-tagged.  This
change adds a `tags_only` flag that can be used to make sure that happens.

Also had to correct a call to `ToBoolPointer()` because it was causing the
plugin to segfault on a null pointer without it.